Parse date into ISO 8601 format instead of seconds since UTC epoch
Fix #14 by not converting to local or UTC date, instead slightly
modifying DateTime strings into easy-to-parse ISO 8601 format.

When a DateTime string follows the EXIF spec, we map some colons to
dashes to conform to ISO 8601.  The other supported format was
ISO 8601, so we don't do anything in this case.